requirements.txt
requests>=2.28
python-dotenv>=1.0
rich>=13.7
